strategy_ide/requirements.txt
# Strategy IDE - Algorithmic Trading Framework
# Python 3.11+

# Alpaca API
alpaca-py>=0.19.0

# Data & indicators (Finnhub for bars; Alpaca for orders only)
requests>=2.28.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
numpy>=1.24.0

# Backtesting (lightweight custom engine included; optional)
# backtrader>=1.9.78

# Charting
plotly>=5.18.0
mplfinance>=0.12.10b0

# Web UI
streamlit>=1.28.0

# Config
python-dotenv>=1.0.0
pytz>=2023.3

# Optimization
hyperopt>=0.2.7

# Optional: type checking
# mypy>=1.5.0
